COMP: Remove signed to unsigned comparison warnings
Use explicit static_casts to compare signed to unsigned values.

+++ Libs/MRML/Core/vtkMRMLStorageNode.cxx
@@ -1582,7 +1582,7 @@ std::string vtkMRMLStorageNode::ClampFileName(const std::string& filename, int e
 
   // Remove extension
   std::string baseName = filename.substr(0, filename.length() - extensionLength);
-  if (baseName.length() <= maxFileNameLength)
+  if (baseName.length() <= static_cast<unsigned int>(maxFileNameLength))
   {
     return filename;
   }
